moving things back to libnavigation-telemetry
