+++ libnavigation-base/build.gradle
@@ -31,9 +31,7 @@ android {
 
 dependencies {
     api dependenciesList.mapboxCore
-    api dependenciesList.mapboxSdkGeoJSON
     api dependenciesList.mapboxSdkDirectionsModels
-    api dependenciesList.mapboxEvents
 
     implementation dependenciesList.supportAnnotation
