+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/telemetry/telemetryevents/MetricEx.kt
@@ -1,10 +1,10 @@
 package com.mapbox.navigation.core.telemetry.telemetryevents
 
 import com.mapbox.android.telemetry.Event
-import com.mapbox.navigation.base.metrics.DirectionsMetrics
-import com.mapbox.navigation.base.metrics.MetricEvent
-import com.mapbox.navigation.base.metrics.NavigationAppUserTurnstileEvent
-import com.mapbox.navigation.base.metrics.NavigationMetrics
+import com.mapbox.navigation.metrics.internal.NavigationAppUserTurnstileEvent
+import com.mapbox.navigation.metrics.internal.utils.extentions.DirectionsMetrics
+import com.mapbox.navigation.metrics.internal.utils.extentions.MetricEvent
+import com.mapbox.navigation.metrics.internal.utils.extentions.NavigationMetrics
 
 fun MetricEvent.toTelemetryEvent(): Event? =
     when (metricName) {
