+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/telemetry/telemetryevents/TelemetryFasterRoute.kt
@@ -3,8 +3,8 @@ import android.annotation.SuppressLint
 import android.os.Parcel
 import com.google.gson.Gson
 import com.mapbox.android.telemetry.Event
-import com.mapbox.navigation.base.metrics.MetricEvent
-import com.mapbox.navigation.base.metrics.NavigationMetrics
+import com.mapbox.navigation.metrics.internal.utils.extentions.MetricEvent
+import com.mapbox.navigation.metrics.internal.utils.extentions.NavigationMetrics
 
 /**
  * Documentation is here [https://paper.dropbox.com/doc/Navigation-Telemetry-Events-V1--AuUz~~~rEVK7iNB3dQ4_tF97Ag-iid3ZImnt4dsW7Z6zC3Lc]
@@ -17,7 +17,7 @@ class TelemetryFasterRoute(
     val newDurationRemaining: Int = 0,
     val newGeometry: String? = null,
     val step: TelemetryStep? = null,
-    var Metadata: TelemetryMetadata
+    var metadata: TelemetryMetadata
 ) : MetricEvent, Event() {
     val event = "navigation.fasterRoute"
     override fun writeToParcel(dest: Parcel?, flags: Int) {
