+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/MapboxMetricsReporter.kt
@@ -3,10 +3,10 @@ package com.mapbox.navigation.metrics
 import android.content.Context
 import com.google.gson.Gson
 import com.mapbox.android.telemetry.MapboxTelemetry
-import com.mapbox.navigation.base.metrics.MetricEvent
-import com.mapbox.navigation.base.metrics.MetricsObserver
-import com.mapbox.navigation.base.metrics.MetricsReporter
-import com.mapbox.navigation.base.metrics.toTelemetryEvent
+import com.mapbox.navigation.metrics.internal.MetricsObserver
+import com.mapbox.navigation.metrics.internal.MetricsReporter
+import com.mapbox.navigation.metrics.internal.utils.extentions.MetricEvent
+import com.mapbox.navigation.metrics.internal.utils.extentions.toTelemetryEvent
 import com.mapbox.navigation.utils.thread.JobControl
 import com.mapbox.navigation.utils.thread.ThreadController
 import kotlinx.coroutines.cancelChildren
