+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/internal/NavigationAppUserTurnstileEvent.kt
@@ -1,7 +1,9 @@
-package com.mapbox.navigation.base.metrics
+package com.mapbox.navigation.metrics.internal
 
 import com.google.gson.Gson
 import com.mapbox.android.telemetry.AppUserTurnstile
+import com.mapbox.navigation.metrics.internal.utils.extentions.MetricEvent
+import com.mapbox.navigation.metrics.internal.utils.extentions.NavigationMetrics
 
 class NavigationAppUserTurnstileEvent(
     val event: AppUserTurnstile
