+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/internal/utils/extentions/MetricEvent.kt
@@ -1,21 +1,22 @@
-package com.mapbox.navigation.base.metrics
+package com.mapbox.navigation.metrics.internal.utils.extentions
 
 import androidx.annotation.StringDef
 import com.google.gson.Gson
 import com.mapbox.android.telemetry.Event
+import com.mapbox.navigation.metrics.internal.NavigationAppUserTurnstileEvent
 
 interface MetricEvent {
 
     @StringDef(
-        DirectionsMetrics.ROUTE_RETRIEVAL,
-        NavigationMetrics.ARRIVE,
-        NavigationMetrics.CANCEL_SESSION,
-        NavigationMetrics.DEPART,
-        NavigationMetrics.REROUTE,
-        NavigationMetrics.FEEDBACK,
-        NavigationMetrics.INITIAL_GPS,
-        NavigationMetrics.FASTER_ROUTE,
-        NavigationMetrics.APP_USER_TURNSTILE
+            DirectionsMetrics.ROUTE_RETRIEVAL,
+            NavigationMetrics.ARRIVE,
+            NavigationMetrics.CANCEL_SESSION,
+            NavigationMetrics.DEPART,
+            NavigationMetrics.REROUTE,
+            NavigationMetrics.FEEDBACK,
+            NavigationMetrics.INITIAL_GPS,
+            NavigationMetrics.FASTER_ROUTE,
+            NavigationMetrics.APP_USER_TURNSTILE
     )
     annotation class Metric
